Major redesigns completed
- Renamed views and routes for create, read, update, and delete
- Implemented the 3 user roles in accessibility and privileges (admin, user, guest)
- Moved the Create-Record view from sidebar to the Table-of-Graduates view
- Made controllers for Table of Graduates, Job Vacancies, and Via Google Sheets

+++ tesda_etrak/routes/web.php
@@ -1,8 +1,10 @@
 <?php
 
 use App\Http\Controllers\AuthController;
-use App\Http\Controllers\EtrakController;
+use App\Http\Controllers\HomeController;
 use App\Http\Controllers\JobVacanciesController;
+use App\Http\Controllers\TableOfGraduatesController;
+use App\Http\Controllers\ViaGoogleSheetsController;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/laravel', function () {
@@ -22,48 +24,49 @@ Route::middleware('guest')->controller(AuthController::class)->group(function ()
 
 Route::post('/logout', [AuthController::class, 'logout'])->name('logout')->middleware('auth');
 
-Route::middleware(['auth', 'role:user'])->controller(EtrakController::class)->group(function () {
-    Route::get('/', 'index')->name('index');
-    Route::get('/dashboard', 'dashboard')->name('dashboard');
-    Route::get('/view-records', 'view_records')->name('view-records');
-    Route::get('/view-records/get', 'search_graduates')->name('search-graduates');
-    Route::get('/record-details/{graduate}', 'view_details')->name('view.details');
-});
-
-Route::controller(EtrakController::class)->group(function () {
-    Route::get('/', 'index')->name('index');
-    Route::get('/dashboard', 'dashboard')->name('dashboard');
-    Route::get('/view-records', 'view_records')->name('view-records');
-    Route::get('/view-records/get', 'search_graduates')->name('search-graduates');
-    Route::get('/record-details/{graduate}', 'view_details')->name('view.details');
-});
-
-Route::controller(JobVacanciesController::class)->group(function () {
-    Route::get('/job-vacancies', 'index')->name('view.vacancies');
-    Route::get('/job-vacancies/get', 'search_vacancies')->name('search-vacancies');
-});
-
-Route::middleware(['auth', 'role:admin,super-admin'])->controller(EtrakController::class)->group(function () {
-    Route::get('/admin', 'index')->name('admin.index');
+Route::middleware(['auth', 'role:admin'])->controller(HomeController::class)->group(function () {
+    Route::get('/admin', 'index')->name('admin.home');
     Route::get('/admin/dashboard', 'dashboard')->name('admin.dashboard');
-    Route::get('/admin/view-records', 'view_records')->name('admin.view-records');
-    Route::get('/admin/view-records/get', 'search_graduates')->name('admin.search-graduates');
-    Route::get('/admin/create-record', 'view_create')->name('admin.view-create');
-    Route::get('/admin/record-details/{graduate}', 'view_details')->name('admin.view-details');
-    Route::get('/admin/update-record/{graduate}', 'view_update')->name('admin.view-update');
-    Route::post('/admin/create-record', 'create')->name('create');
-    Route::put('/admin/update-record/{graduate}', 'update')->name('update');
-    Route::delete('/admin/record-details/{graduate}', 'delete')->name('delete');
-    Route::delete('/admin/view-records', 'delete_all')->name('delete-all');
-    Route::get('/admin/google-sheets-data', 'view_sheets_data')->name('admin.view-sheets-data');
-    Route::get('/admin/google-sheets-data/import-data', 'import_data')->name('import.data');
-    Route::get('/admin/google-sheets-data/export-data', 'export_data')->name('export.data');
-    Route::get('/admin/google-sheets-data/logs', 'display_logs')->name('display.log');
-    Route::post('/admin/google-sheets-data/logs/clear', 'clear_logs')->name('clear.logs');
+});
+
+Route::controller(HomeController::class)->group(function () {
+    Route::get('/', 'index')->name('home');
+    Route::get('/dashboard', 'dashboard')->name('dashboard');
 });
 
 Route::middleware(['auth', 'role:admin'])->controller(JobVacanciesController::class)->group(function () {
     Route::get('/admin/job-vacancies', 'index')->name('admin.view-vacancies');
     Route::get('/admin/job-vacancies/get', 'search_vacancies')->name('admin.search-vacancies');
     Route::get('/admin/job-vacancies/import-data', 'import_data')->name('import.vacancies.data');
+});
+
+Route::get('/job-vacancies', [JobVacanciesController::class, 'index'])->name('job-vacancies');
+
+Route::get('/job-vacancies/search', [JobVacanciesController::class, 'search_vacancies'])->name('search-vacancies')->middleware('auth', 'role:user');
+
+Route::middleware(['auth', 'role:admin'])->controller(TableOfGraduatesController::class)->group(function () {
+    Route::get('/admin/table-of-graduates', 'index')->name('admin.table-of-graduates');
+    Route::get('/admin/table-of-graduates/search', 'search_graduates')->name('admin.search-graduates');
+    Route::get('/admin/table-of-graduates/create-record', 'create_view')->name('admin.create-record.view');
+    Route::get('/admin/table-of-graduates/record-details/{graduate}', 'read')->name('admin.record-details');
+    Route::get('/admin/table-of-graduates/update-record/{graduate}', 'update_view')->name('admin.update-record.view');
+    Route::post('/admin/table-of-graduates/create-record', 'create')->name('admin.create-record');
+    Route::put('/admin/table-of-graduates/update-record/{graduate}', 'update')->name('admin.update-record');
+    Route::delete('/admin/table-of-graduates/record-details/{graduate}', 'delete')->name('admin.delete-record');
+    Route::delete('/admin/table-of-graduates', 'truncate')->name('admin.truncate-graduates');
+});
+
+Route::get('/table-of-graduates', [TableOfGraduatesController::class, 'index'])->name('table-of-graduates');
+
+Route::middleware(['auth', 'role:user'])->controller(TableOfGraduatesController::class)->group(function () {
+    Route::get('/table-of-graduates/search', 'search_graduates')->name('search-graduates');
+    Route::get('/table-of-graduates/record-details/{graduate}', 'read')->name('record-details');
+});
+
+Route::middleware(['auth', 'role:admin'])->controller(ViaGoogleSheetsController::class)->group(function () {
+    Route::get('/admin/via-google-sheets', 'index')->name('admin.via-google-sheets');
+    Route::get('/admin/via-google-sheets/import-graduate-sheet', 'import_sheet')->name('admin.import-graduate');
+    Route::get('/admin/via-google-sheets/export-graduate-data', 'export_data')->name('admin.export-graduate');
+    Route::get('/admin/via-google-sheets/logs', 'display_logs')->name('admin.display.log');
+    Route::post('/admin/google-sheets-data/logs/clear', 'clear_logs')->name('admin.clear.logs');
 });
